+++ src/day2/range/SameLengthRange.kt
@@ -1,6 +1,6 @@
 package day2.range
 
-class SameLengthRange(start: Int, end: Int) : Range(start = start, end = end) {
+class SameLengthRange(start: Long, end: Long) : Range(start = start, end = end) {
     init {
         require(startDigitCount == endDigitCount) { "start and end must be of the same length" }
     }
